requirements.txt
# Core dependencies
numpy>=1.26.0
scipy>=1.11.3
matplotlib>=3.8.0
gymnasium>=0.28.1
pyyaml>=6.0.1

# Reinforcement learning
stable-baselines3>=2.1.0
torch>=2.0.1
tensorboard>=2.14.0

# Visualization
pygame>=2.5.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6

# Optional for video recording
ffmpeg-python>=0.2.0

# Utils
tqdm>=4.66.1
pandas>=2.1.1
seaborn>=0.13.0
pillow>=10.0.1 